+++ assets/jupyter/jupyter_notebook_config.py
@@ -8,7 +8,6 @@ c.NotebookApp.notebook_dir              = '{{ fsleyes_nbserver_dir }}'
 c.NotebookApp.extra_static_paths        = ['{{ fsleyes_nbserver_static_dir }}']
 c.NotebookApp.answer_yes                = True
 c.NotebookApp.extra_nbextensions_path   = ['{{ fsleyes_nbextension_dir }}']
-c.NotebookApp.open_browser              = False
 c.NotebookApp.kernel_manager_class      = 'fsleyes_kernelmanager.FSLeyesNotebookKernelManager'
 c.NotebookApp.kernel_spec_manager_class = 'fsleyes_kernelspecmanager.FSLeyesKernelSpecManager'